+++ NEWS.md
@@ -1,3 +1,3 @@
 # digitTests 0.1.0
 
-- This is `digitTests` version `0.1.0` with functions `distr.test()` (test digits against a reference, i.e. benford, distribution) and `rv.test()` (repeated values test).
+- This is `digitTests` version `0.1.0` with functions `distr.test()`, `distr.btest()` (test first, first two, or last digits against a reference, i.e. benford or uniform, distribution) and `rv.test()` (test if data contain an excessive amount of repeated values).